Ignore ruff errors in resenet_density (#22)
While it would be good to resolve these errors, we are not currently
developing `src/resenet_density/`, and my goal is to get us to a point
where CI passes so we catch new errors. If we begin iterating on
`resenet_density`, we can go through and address formatting at that
time.